Array and array index typechecking complete
This shone light on some flaws in the design. The requirement for an introduction of the "Any" type required modifying every other type, which is problematic. I'm not sure what I can do to improve this.

The array assignment code is extremely messy and contains much potentially fragile typecasting. A better system should be devised, but I'm not sure what yet. Of special note is the difficulty in transforming the array's Eq size constraint to something we can use to typecheck the array bounds. Very problematic!

+++ Compiler/TypeChecker/Constraints/Lt.cs
@@ -6,9 +6,9 @@ using System.Threading.Tasks;
 
 namespace Speedycloud.Compiler.TypeChecker.Constraints {
     public class Lt : ITypeConstraint {
-        public int Num { get; private set; }
+        public decimal Num { get; private set; }
 
-        public Lt(int num) {
+        public Lt(decimal num) {
             Num = num;
         }
